+++ modules/admin-ui-picard/app/src/components/events/partials/ModalTabsAndPages/DetailsMetadataTab.js
@@ -8,7 +8,8 @@ import RenderMultiField from "../../../shared/wizard/RenderMultiField";
 import RenderField from "../../../shared/wizard/RenderField";
 import {getUserInformation} from "../../../../selectors/userInfoSelectors";
 import {connect} from "react-redux";
-import {hasAccess} from "../../../../utils/utils";
+import {hasAccess, isJson} from "../../../../utils/utils";
+import {getMetadataCollectionFieldName} from "../../../../utils/resourceUtils";
 
 /**
  * This component renders metadata details of a certain event or series
@@ -73,7 +74,15 @@ const DetailsMetadataTab = ({ metadataFields, updateResource, resourceId, header
                                                     </td>
                                                     {field.readOnly ? (
                                                         // non-editable field if readOnly is set
-                                                        <td>{field.value}</td>
+                                                        (!!field.collection && field.collection.length !== 0) ? (
+                                                            <td>
+                                                                {isJson(getMetadataCollectionFieldName(field, field)) ?
+                                                                    (t(JSON.parse(getMetadataCollectionFieldName(field, field)).label)) :
+                                                                    (t(getMetadataCollectionFieldName(field, field)))}
+                                                            </td>
+                                                        ) : (
+                                                            <td>{field.value}</td>
+                                                        )
                                                     ) : (
                                                         <td className="editable ng-isolated-scope">
                                                             {/* Render single value or multi value editable input */}
